implement custom hook for fps tracking
